requirements.txt
ffmpeg
giphy_client
discord.py[voice]
emoji
youtube_dl
python-dotenv
twilio
opuslib